[Refector] 랜덤 이름 함수를 유틸 폴더로 이동, x버튼 이동
